Update Mainpage.html
This will be the mainpage of our application. if you click on the "create the project" button, the form page will pop up. The index.html in this folder will be our project page. whenever the user clicks on the project from their project list index.html will open with all the sidebar properties.